deployment_service.sh does not launch
The deployment_service.sh script does not appear to do anything when launched. The code below the "Deploy the function service" line tests whether the service exists, and if it does, prompts if you want to reinstall/overwrite. However in the event that the service is not installed, the code will never run the install_service function.

It looks like that portion of the code is missing an else statement to run the function.

Added the lines:
else
    echo "Installing UCtronics display service"
    install_service